Fix fixes... *sigh* (js@codeartist.org--2004-public/mel-base--dev--0.5--patch-3) js@codeartist.org--2004-public/mel-base--dev--0.5--patch-3 Keywords:
I missed the SBCL reader conditional in the system
definition. Added it now.

There seem to be problems with using SLOT-BOUNDP
in CMUCLs lacking CLOS integration for conditions.
I changed it to use an initform of NIL to represent
the absent slot-value.

Implemented strict matching of month names in date
parsing. Fixed the bug were "Jan" would be matched
wrong too.